Removed bounds checking (see description).
For some reason our code does not compile when we enable bounds checking. Even after removing all of our code and just having SDL's stuff it still would not run correctly. This could potentially mean it's an issue with TCC and SDL? So for the time being we will just disable bounds checking and stick to just having debug info in our debug builds.